pyproject.toml
[tool.poetry]
name = "java-luchshe"
version = "0.1.0"
description = ""
authors = ["Angelina Stiukhina <a.stiukhina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
pydantic = "^2.9.2"
databases = "^0.9.0"
pydantic-settings = "^2.6.0"
greenlet = "^3.1.1"
alembic = "^1.13.3"
tenacity = "^9.0.0"
psycopg2 = "^2.9.10"
bcrypt = "^4.2.1"
python-multipart = "^0.0.19"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
